Wait after login in SSH connection
We need to wait a second before running anny command after SSH
connection. This will ensure we waited for the welcome message and we
can login using sudo.
This patch is a hack, due to the fact we are not able to setup welcome
message in the SSH server from client.